pyproject.toml
[project]
name = "gas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "matplotlib>=3.10.7",
]
